Revert to pre-alpha state
