+++ microgp4/user_messages/exception.py
@@ -32,9 +32,8 @@ __all__ = ['MicroGPException', 'MicroGPError', 'InvalidIndividual', 'GeneticOper
 
 class MicroGPException(Exception):
     """Base class for exceptions in MicroGP4."""
-    def __init__(self, *args: object) -> None:
-        super().__init__(*args)
-        
+
+
 class MicroGPError(MicroGPException):
     """Exception for a serious error in MicroGP4."""
