+++ test/test_global_symbols.py
